requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.0
django-debug-toolbar==2.0
django-shortuuidfield==0.1.3
djangorestframework==3.10.2
idna==2.8
mysqlclient==1.4.2.post1
Pillow==6.1.0
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.1
qiniu==7.2.6
redis==3.2.1
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
